Magdy dev

+++ app/Http/Controllers/Dashboard/BrancheController.php
@@ -3,13 +3,12 @@
 namespace App\Http\Controllers\Dashboard;
 
 use App\Branch;
-use App\Plan;
 use App\Http\Controllers\Controller;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Gate;
 
-class PlansController extends Controller
+class BrancheController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -19,18 +18,22 @@ class PlansController extends Controller
     public function __construct()
     {
         $this->middleware('auth');
-
     }
 
     public function index(Request $request)
     {
-        if (! Gate::allows('show_plans')) {
+        if (!Gate::allows('show_branches')) {
             return abort(401);
         }
-        // $plans  = Plan::get();
-        // return $plans;
-        $plans  = Plan::all();
-        return view('plans.index', compact('plans'));
+        try {
+            if ($request->type == 'table') {
+                return view('branches.index_table', ['branches' => Branch::all()]);
+            }
+            $branches = Branch::withDepth()->with('ancestors')->get()->toTree();
+            return view('branches.index', compact('branches'));
+        } catch (Exception $exp) {
+            Branch::fixTree();
+        }
     }
 
     /**
@@ -63,13 +66,13 @@ class PlansController extends Controller
         $branch = new Branch($data);
         $branch->save();
         // return $request->parent_id;
-        if($request->parent_id ){
+        if ($request->parent_id) {
             // $parent = Branch::find($request->parent_id);
             $branch->parent_id = $request->parent_id;
             $branch->save();
         }
 
-        return back()->with('success' , 'تم الحفظ بنجاح' );
+        return back()->with('success', 'تم الحفظ بنجاح');
     }
 
     /**
@@ -91,12 +94,12 @@ class PlansController extends Controller
      */
     public function edit($id)
     {
-        if (! Gate::allows('edit_plan')) {
+        if (!Gate::allows('edit_branch')) {
             return abort(401);
         }
         //
         $branch = Branch::find($id);
-        return view('branches.edit' , ['branch' => $branch]);
+        return view('branches.edit', ['branch' => $branch]);
     }
 
     /**
@@ -108,12 +111,12 @@ class PlansController extends Controller
      */
     public function update(Request $request, $id)
     {
-        try{
+        try {
             $branch = Branch::findOrFail($id);
             //update in db
             $branch->update($request->all());
-            return redirect()->route('branches.index',['type' => 'table'])->with(['success' => 'تم تحديث الفرع بنجاح']);
-        }catch(\Exception $ex){
+            return redirect()->route('branches.index', ['type' => 'table'])->with(['success' => 'تم تحديث الفرع بنجاح']);
+        } catch (\Exception $ex) {
             return back()->with(['error' => 'هناك خطأ برجاء المحاولة ثانيا']);
         }
     }
@@ -126,14 +129,16 @@ class PlansController extends Controller
      */
     public function destroy($id)
     {
-        try{
+        try {
             $branch = Branch::findOrFail($id);
-            //delete in db
-            $branch ->delete();
-            // Branch::fixTree();
-            return back()->with(['success' => 'تم حذف الفرع بنجاح']);
-
-        }catch(\Exception $ex){
+            if ($branch->can_delete) {
+                //delete in db
+                $branch->delete();
+                return back()->with(['success' => 'تم حذف الفرع بنجاح']);
+            } else {
+                return back()->with(['error' => 'هذا الفرع يوجد عليه موظفين لا يمكن مسحه']);
+            }
+        } catch (\Exception $ex) {
             return back()->with(['error' => 'هناك خطأ برجاء المحاولة ثانيا']);
         }
     }
